pyproject.toml
[tool.poetry]
name = "neural-networks"
version = "0.1.0"
description = ""
authors = ["loganpowell <loganp@tepper.cmu.edu>"]
readme = "README.md"

[tool.poetry.dependencies]
azure-cosmos = "^4.6.0"
azure-identity = "^1.15.0"
azure-keyvault-secrets = "^4.8.0"
azure-search-documents = "^11.4.0"
azure-storage-blob = "^12.19.1"
beautifulsoup4 = "^4.12.3"
datasets = "^2.18.0"
dohq-artifactory = "^0.10.0"
glom = "^23.5.0"
gremlinpython = "^3.7.1"
markdownify = "^0.11.6"
openai = "^1.13.3"
pathlib-abc = "^0.2.0"
python = ">=3.10,<3.12"
requests = "^2.31.0"
scikit-learn = "^1.4.1.post1"
seaborn = "^0.13.2"
tiktoken = "^0.6.0"
torch = "^2.2.1"
torchvision = "^0.17.1"
xmltodict = "^0.13.0"

[tool.poetry.group.dev.dependencies]
async-timeout = "^4.0.3"
ipykernel = "^6.29.3"
ipywidgets = "^8.1.2"
matplotlib = "^3.8.3"
numpy = "^1.26.4"
pandas = "^2.2.1"
python-dotenv = "^1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
